[KangJingA/integrate 1901] merge data mapping flow to white rabbit (#65)
* create data mapping tasks

* update white rabbit flows with data mapping

* update ui

* remove unneeded endpoints

* remove data mapping flow

* fix link table button disabled after clicking back

* move task

* fix button
