--- flows/data_mapping/package.json
@@ -1,75 +1,0 @@
-{
-  "name": "@data2evidence/data-mapping-flow",
-  "version": "0.0.99",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "build": "docker build ../ -f ./Dockerfile -t d2e-flows-local-data-mapping --build-arg GITHUB_PAT=${GITHUB_PAT} --platform=linux/amd64"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/data2evidence/d2e-flows.git"
-  },
-  "author": "",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/data2evidence/d2e-flows/issues"
-  },
-  "homepage": "https://github.com/data2evidence/d2e-flows#readme",
-  "trex": {
-    "flow": {
-      "image": "ghcr.io/ohdsi/d2e/flow-data-mapping",
-      "flows": [
-        {
-          "name": "data_mapping_plugin",
-          "type": "perseus",
-          "entrypoint": "flows.data_mapping_plugin.flow.data_mapping_plugin",
-          "image": "ghcr.io/ohdsi/d2e/flow-data-mapping:develop",
-          "image_pull_policy": "Never",
-          "tags": [
-            "data_mapping"
-          ],
-          "parameter_openapi_schema": {
-            "title": "Parameters",
-            "type": "object",
-            "properties": {
-              "options": {
-                "$ref": "#/definitions/PerseusRequestType",
-                "position": 0,
-                "title": "options"
-              }
-            },
-            "required": [
-              "options"
-            ],
-            "definitions": {
-              "PerseusRequestType": {
-                "properties": {
-                  "url": {
-                    "title": "Url",
-                    "type": "string"
-                  },
-                  "headers": {
-                    "title": "Headers",
-                    "type": "object"
-                  },
-                  "data": {
-                    "title": "Data",
-                    "type": "object"
-                  }
-                },
-                "required": [
-                  "url",
-                  "headers",
-                  "data"
-                ],
-                "title": "PerseusRequestType",
-                "type": "object"
-              }
-            }
-          }
-        }
-      ]
-    }
-  }
-}
