--- flows/data_mapping/requirements_perseus.txt
@@ -1,9 +1,0 @@
-prefect[shell, docker]==3.0.3
-pydantic==2.10.6
-python-dotenv==1.0.0
-jwt==1.3.1
-xmltodict==0.14.2
-pandas==2.2.0
-python-calamine==0.3.1
-pandasql==0.7.3
-sqlalchemy==2.0.38
