+++ ui/apps/mapping/src/components/ScanProgressDialog/ScanProgressDialog.scss
@@ -9,27 +9,6 @@
     align-items: center;
   }
 
-  .scan-progress-dialog__loader {
-    margin-left: auto;
-    display: flex;
-    align-items: center;
-    
-    :first-child {
-      width: 24px !important;
-      height: 24px !important;
-      
-      svg {
-        width: 24px !important;
-        height: 24px !important;
-      }
-      
-      .MuiCircularProgress-root {
-        width: 24px !important;
-        height: 24px !important;
-      }
-    }
-  }
-
   .scan-progress-dialog__content {
     padding: 1em;
     display: flex;
